Remove `bin/run-tests` and use testem, as we already do for browser-specific tests
bin/run-tests was painful to maintain and did not have any output per-test (it had been lost).
