+++ .gitignore
@@ -46,7 +46,6 @@ npm-debug.log
 *.tar.gz
 *.log
 /.vscode
-.puppeteer-cache
 
 # These are automatically generated by our build process.
 # TODO: make that *fully* true: The root types/stable directory is *not*
